修正變數類型

+++ src/main/java/fin/Star.java
@@ -7,16 +7,15 @@ public class Star extends PApplet {
     int width = 1900;
     int height = 1050;
     Ball[] ball = new Ball[num];
-    double[] ax = new double[num];
-    double[] ay = new double[num];
+    float[] ax = new float[num];
+    float[] ay = new float[num];
     boolean back = false;
-    double magnetism = 10.0; 
-    double radius = 1; 
+    float magnetism = 10.0f; 
+    float radius = 1; 
     float gspeed = 0.95f; 
 
     public void settings() {
-        // 
-        size(1920, 1080);					//size(width, height)，設定視窗的長寬像素點的個數
+        size(1920, 1080); //size(width, height)，設定視窗的長寬像素點的個數
         for (int i = 0; i < num; i++) {
             ball[i] = new Ball(random(width),random(height),0,0);
         }
@@ -26,24 +25,23 @@ public class Star extends PApplet {
     	background_setting();
 
         for (int i = 0; i < num; i++) {
-            double distance = dist((float) mouseX, (float) mouseY, ball[i].get_x(), ball[i].get_y()); //找到兩點之間距離
+            float distance = dist( mouseX, mouseY, ball[i].get_x(), ball[i].get_y()); //找到兩點之間距離
             //加速度與距離吸引中心的距離的平方成反比。
             if (distance > 3) { //如果距離鼠標太遠
                 ax[i] = magnetism * (mouseX - ball[i].get_x()) / (distance * distance);
                 ay[i] = magnetism * (mouseY - ball[i].get_y()) / (distance * distance);
             }
-            ball[i].acceleration_vx((float)ax[i]);//增加速度
-            ball[i].acceleration_vy((float)ay[i]);
+            ball[i].acceleration_vx(ax[i]);//增加速度
+            ball[i].acceleration_vy(ay[i]);
             ball[i].add_gspeed(gspeed);//重力
-
             ball[i].add_x_v();//新增加速度
 
             float speed = dist(0, 0, ball[i].get_vx(), ball[i].get_vy()); //從向量中找出速度
-            double r = map(speed, 0, 5, 0, 255); //根據速度計算顏色
-            double g = map(speed, 0, 5, 64, 255);
-            double b = map(speed, 0, 5, 128, 255);
-            fill((float) r, (float) g, (float) b, 32);
-            ellipse(ball[i].get_x(), ball[i].get_y(), (float) radius, (float) radius);
+            float r = map(speed, 0, 5, 0, 255); //根據速度計算顏色
+            float g = map(speed, 0, 5, 64, 255);
+            float b = map(speed, 0, 5, 128, 255);
+            fill(r, g, b, 32);
+            ellipse(ball[i].get_x(), ball[i].get_y(), radius, radius);
         }
     }
     void background_setting() {
